pyproject.toml
[tool.poetry]
name = "meltanolabs-tap-github"
version = "1.11.1"
description = "Singer tap for GitHub, built with the Singer SDK."
authors = ["Meltano and Meltano Community <hello@meltano.com>"]
maintainers = [
	"Meltano and Meltano Community <hello@meltano.com>",
	"Edgar Ramírez-Mondragón <edgarrm358@gmail.com>",
]
homepage = "https://github.com/MeltanoLabs/tap-github"
repository = "https://github.com/MeltanoLabs/tap-github"
license = "Apache-2.0"
keywords = ["Meltano", "Singer", "Meltano SDK", "Singer SDK", "ELT", "Github"]
readme = "README.md"
classifiers = [
	"Intended Audience :: Developers",
	"License :: OSI Approved :: Apache Software License",
	"Operating System :: OS Independent",
	"Programming Language :: Python :: 3.9",
	"Programming Language :: Python :: 3.10",
	"Programming Language :: Python :: 3.11",
	"Programming Language :: Python :: 3.12",
	"Programming Language :: Python :: 3.13",
	"Programming Language :: Python :: Implementation :: CPython",
	"Typing :: Typed",
]
packages = [
	{ include = "tap_github", format = ["sdist", "wheel"] }
]

[tool.poetry.urls]
"Issue Tracker" = "https://github.com/MeltanoLabs/tap-github/issues"

[tool.poetry.dependencies]
beautifulsoup4 = "~=4.12.0"
cryptography = { version = "~=44.0.0", python = ">3.9.0,<3.9.1 || >3.9.1" }
nested-lookup = "~=0.2.25"
PyJWT = "2.10.1"
python = ">=3.9"
python-dateutil = "~=2.9"
requests = "~=2.32.3"
# For local SDK dev:
# singer-sdk = {path = "../singer-sdk", develop = true}
singer-sdk = "~=0.44.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.3.1"
requests-cache = ">=1.0.1"
types-beautifulsoup4 = ">=4.12.0"
types-python-dateutil = "~=2.9.0"
types-requests = ">=2.30.0"
types-simplejson = "~=3.19.0"

[[tool.mypy.overrides]]
module = ["backoff"]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-github = 'tap_github.tap:cli'

[tool.pytest.ini_options]
markers = [
	"repo_list: mark a test as using a list of repos in config",
	"username_list: mark a test as using a list of usernames in config",
]

[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
ignore = []
select = [
	"F",    # Pyflakes
	"E",    # pycodestyle (errors)
	"W",    # pycodestyle (warnings)
	"I",    # isort
	"N",    # pep8-naming
	"UP",   # pyupgrade
	"YTT",  # flake8-2020
	"ANN",  # flake8-annotations
	"B",    # flake8-bugbear
	"A",    # flake8-builtins
	"C4",   # flake8-comprehensions
	"DTZ",  # flake8-datetimez
	"FA",   # flake8-future-annotations
	"SIM",  # flake8-simplify
	"TC",   # flake8-type-checking
	"PERF", # Perflint
	"FURB", # refurb
	"RUF",  # Ruff-specific rules
]

[tool.ruff.lint.per-file-ignores]
"tap_github/tests/*" = ["ANN"]
